% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/b_linear.R
\name{b_linear}
\alias{b_linear}
\title{b_linear}
\usage{
b_linear(x, y, s, p_mu_a = NULL, p_sigma_a = NULL, p_mu_b = NULL,
  p_sigma_b = NULL, p_mu_s = NULL, p_sigma_s = NULL, warmup = 2000,
  iter = 3000, chains = 4, control = NULL)
}
\arguments{
\item{x}{a vector containting index of sequence (time).}

\item{y}{a vector containting subjet's responses.}

\item{s}{a vector contaiting subject indexes. Starting index should be 1, and the largest subject index equals m (number of subjects).}

\item{p_mu_a}{Prior for the mu_a (mean intercept) parameter (default = NULL/flat improper prior).}

\item{p_sigma_a}{Prior for the sigma_a (mu_a variance) parameter (default = NULL/flat improper prior).}

\item{p_mu_b}{Prior for the mu_b (mean slope) parameter (default = NULL/flat improper prior).}

\item{p_sigma_b}{Prior for the sigma_b (mu_b variance) parameter (default = NULL/flat improper prior).}

\item{p_mu_s}{Prior for the mu_s (variance) parameter (default = NULL/flat improper prior).}

\item{p_sigma_s}{Prior for the sigma_s (mu_s variance) parameter (default = NULL/flat improper prior).}

\item{warmup}{Integer specifying the number of warmup iterations per chain (default = 2000).}

\item{iter}{Integer specifying the number of iterations (including warmup, default = 3000).}

\item{chains}{Integer specifying the number of parallel chains (default = 4).}

\item{control}{A named list of parameters to control the sampler's behavior (default = NULL).}
}
\value{
An object of class `linear_class`.
}
\description{
Bayesian model for fitting a linear normal model to data.
}
