% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/b_prior.R
\docType{class}
\name{b_prior-class}
\alias{b_prior-class}
\alias{b_prior}
\title{b_prior}
\description{
An S4 class for defining priors for Bayesian models.
}
\section{Slots}{

\describe{
\item{\code{family}}{Prior family - \"uniform\", \"normal\", \"gamma\" or \"beta\".}

\item{\code{pars}}{Parameters of the prior - a vector of two numerical values.}
}}

