% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/b_prior.R
\name{get_prior_id}
\alias{get_prior_id}
\alias{get_prior_id,b_prior-method}
\alias{get_prior_id_b_prior}
\title{get_prior_id}
\usage{
get_prior_id(object)

\S4method{get_prior_id}{b_prior}(object)
}
\arguments{
\item{object}{b_prior object.}
}
\description{
\code{get_prior_id} returns an integer id of prior's family (1 = uniform, 2 = normal, 3 = gamma, 4 = beta).
}
