% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/b_reaction_times.R
\name{b_reaction_times}
\alias{b_reaction_times}
\title{Bayesian model for comparing reaction times and test success rate.}
\usage{
b_reaction_times(n, m, rt, r, s)
}
\arguments{
\item{n}{total number of measurements.}

\item{m}{number of participating subjects.}

\item{rt}{a vector containing reaction times for each measurement.}

\item{r}{a vector containting test results (0 - test was not solved successfully, 1 - test was solved succesfully).}

\item{s}{a vector contaiting subject indexes. Starting index should be 1, and the largest subject index equals m (number of subjects).}
}
\value{
An object of class `rt_results_class`.
}
\description{
Bayesian model for comparing reaction times and test success rate.
}
