% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/b_results.R
\docType{class}
\name{b_results-class}
\alias{b_results-class}
\title{b_results}
\description{
Parent S4 class for declaring shared function generics.

\strong{Functions}

summary(`success_rate_class`): prints summary of the fit.

print(`success_rate_class`): prints a more detailed summary of the fit.

show(`success_rate_class`): prints a more detailed summary of the fit.
}
