% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/b_results.R
\name{compare_samples}
\alias{compare_samples}
\title{compare_samples}
\usage{
compare_samples(object, ...)
}
\arguments{
\item{object}{S4 class object from EasyBayes library.}

\item{...}{see documentation for specific class for the description of available parameters, e.g. ?compare_ttest or ?compare_linear.}
}
\description{
\code{compare_samples} prints difference between two groups.
}
