% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/b_results.R
\name{get_parameters}
\alias{get_parameters}
\title{get_parameters}
\usage{
get_parameters(object)
}
\arguments{
\item{object}{S4 class object from bayes4psy library.}
}
\description{
\code{get_parameters} returns a dataframe with values of fitted parameters.
}
