% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/b_results.R
\name{get_samples}
\alias{get_samples}
\title{get_samples}
\usage{
get_samples(object)
}
\arguments{
\item{object}{S4 class object from bayes4psy library.}
}
\description{
\code{get_samples} returns a dataframe with values of fitted parameters.
}
