% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/b_results.R
\name{get_subject_parameters}
\alias{get_subject_parameters}
\title{get_subject_parameters}
\usage{
get_subject_parameters(object)
}
\arguments{
\item{object}{S4 class object from bayes4psy library.}
}
\description{
\code{get_subject_parameters} returns a dataframe with values of fitted parameters for each subject in the hierarchical model.
}
