% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/b_results.R
\name{get_subject_samples}
\alias{get_subject_samples}
\title{get_subject_samples}
\usage{
get_subject_samples(object)
}
\arguments{
\item{object}{S4 class object from EasyBayes library.}
}
\description{
\code{get_subject_samples} returns a dataframe with values of fitted parameters for each subject in the hierarchical model.
}
