% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/b_results.R
\name{plot_difference}
\alias{plot_difference}
\title{plot_difference}
\usage{
plot_difference(object, ...)
}
\arguments{
\item{object}{S4 class object from EasyBayes library.}

\item{...}{see documentation for specific class for the description of available parameters, e.g. ?plot_difference_ttest or ?plot_difference_linear.}
}
\description{
\code{plot_difference} plots difference between two groups.
}
