% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/b_results.R
\name{plot_distributions_difference}
\alias{plot_distributions_difference}
\title{plot_distributions_difference}
\usage{
plot_distributions_difference(object, ...)
}
\description{
plot_distributions_difference
}
