% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/b_results.R
\name{plot_fit}
\alias{plot_fit}
\title{plot_fit}
\usage{
plot_fit(object, ...)
}
\arguments{
\item{object}{S4 class object from EasyBayes library.}
}
\description{
\code{plot_fit} plots fitted model against the data. Use this function to explore the quality of your fit.
}
