% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/b_results.R
\name{plot_means}
\alias{plot_means}
\title{plot_means}
\usage{
plot_means(object, ...)
}
\arguments{
\item{object}{S4 class object from bayes4psy library.}

\item{...}{see documentation for specific class for the description of available parameters, e.g. ?plot_means_ttest or ?plot_means_linear.}
}
\description{
\code{plot_means} plots means for one or multiple fits.
}
