% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/b_results.R
\name{plot_means_difference}
\alias{plot_means_difference}
\title{plot_means_difference}
\usage{
plot_means_difference(object, ...)
}
\arguments{
\item{object}{S4 class object from bayes4psy library.}

\item{...}{see documentation for specific class for the description of available parameters, e.g. ?plot_means_difference_ttest or ?plot_means_difference_linear.}
}
\description{
\code{plot_means_difference} plots difference between means of two or multiple fits.
}
