% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/b_results.R
\name{plot_samples}
\alias{plot_samples}
\title{plot_samples}
\usage{
plot_samples(object, ...)
}
\arguments{
\item{object}{S4 class object from bayes4psy library.}

\item{...}{see documentation for specific class for the description of available parameters, e.g. ?plot_samples_ttest or ?plot_samples_linear.}
}
\description{
\code{plot_samples} plots samples for one or two groups.
}
