% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/b_results.R
\name{plot_samples_difference}
\alias{plot_samples_difference}
\title{plot_samples_difference}
\usage{
plot_samples_difference(object, ...)
}
\arguments{
\item{object}{S4 class object from bayes4psy library.}

\item{...}{see documentation for specific class for the description of available parameters, e.g. ?plot_samples_difference_ttest or ?plot_samples_difference_linear.}
}
\description{
\code{plot_samples_difference} plots difference between two groups.
}
