% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/b_results.R
\name{compare}
\alias{compare}
\alias{plot_difference}
\alias{plot_samples}
\alias{compare_distributions}
\alias{plot_distributions}
\alias{plot_distributions_difference}
\alias{plot_fit}
\alias{plot_trace}
\title{compare}
\usage{
compare(object, ...)

plot_difference(object, ...)

plot_samples(object, ...)

compare_distributions(object, ...)

plot_distributions(object, ...)

plot_distributions_difference(object, ...)

plot_fit(object)

plot_trace(object)
}
\arguments{
\item{object}{S4 class object from EasyBayes library.}

\item{...}{see documentation for specific class for the description of available parameters.}

\item{object}{S4 class object from EasyBayes library.}

\item{...}{see documentation for specific class for the description of available parameters.}

\item{object}{S4 class object from EasyBayes library.}

\item{...}{see documentation for specific class for the description of available parameters.}

\item{object}{S4 class object from EasyBayes library.}

\item{...}{see documentation for specific class for the description of available parameters.}

\item{object}{S4 class object from EasyBayes library.}

\item{...}{see documentation for specific class for the description of available parameters.}

\item{object}{S4 class object from EasyBayes library.}

\item{...}{see documentation for specific class for the description of available parameters.}

\item{object}{S4 class object from EasyBayes library.}

\item{object}{S4 class object from EasyBayes library.}
}
\description{
\code{compare} prints difference between two groups.

\code{plot_difference} plots difference between two groups.

\code{plot_samples} plots samples for one or two groups.

\code{compare_distributions} draws samples from distribution of the first group and compares them against samples drawn from the distribution of the second group,

\code{plot_distributions} visualizes distributions underlying tested groups.

\code{plot_distributions_difference} a visualization of the difference between the distribution of the first group and the distribution or a constant value for the second group.

\code{plot_fit} plots fitted model against the data. Use this function to explore the quality of your fit.

\code{plot_trace} traceplot for main fitted model parameters.
}
