% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/b_success_rate.R
\name{b_success_rate}
\alias{b_success_rate}
\title{b_success_rate}
\usage{
b_success_rate(n, m, r, s)
}
\arguments{
\item{n}{total number of measurements.}

\item{m}{number of participating subjects.}

\item{r}{a vector containting test results (0 - test was not solved successfully, 1 - test was solved succesfully).}

\item{s}{a vector contaiting subject indexes. Starting index should be 1, and the largest subject index equals m (number of subjects).}
}
\value{
An object of class `success_rate_class`.
}
\description{
Bayesian model for comparing reaction times and test success rate.
}
