% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/b_ttest.R
\name{b_ttest}
\alias{b_ttest}
\title{b_ttest}
\usage{
b_ttest(data, p_mu = NULL, p_sigma = NULL, warmup = 9000,
  iter = 10000, chains = 4, control = NULL)
}
\arguments{
\item{data}{Numeric vector of values on which the fit will be based.}

\item{p_mu}{Prior for the mu (mean value) parameter (default = NULL/flat improper prior).}

\item{p_sigma}{Prior for the sigma (variance) parameter (default = NULL/flat improper prior).}

\item{warmup}{Integer specifying the number of warmup iterations per chain (default = 9000).}

\item{iter}{Integer specifying the number of iterations (including warmup, default = 10000).}

\item{chains}{Integer specifying the number of parallel chains (default = 4).}

\item{control}{A named list of parameters to control the sampler's behavior (default = NULL).}
}
\value{
An object of class `ttest_class`.
}
\description{
Bayesian t-test.
}
