% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bayes4psy-package.R
\docType{package}
\name{bayes4psy-package}
\alias{bayes4psy-package}
\alias{bayes4psy}
\title{The 'bayes4psy' package.}
\description{
A user-friendly implementation of Bayesian statistical methods commonly used in social sciences. All used models are pre-compiled, meaning that users only need to call appropriate functions using their data.
}
\references{
Stan Development Team (NA) - the Stan framework and RStan interface.
John Kruschke - mcmc_hdi function
Rasmus Bååth - Easy Bayesian Bootstrap in R
}
