% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/color_class.R
\docType{methods}
\name{compare,color_class-method}
\alias{compare,color_class-method}
\alias{compare_color}
\title{compare}
\usage{
\S4method{compare}{color_class}(object, ...)
}
\arguments{
\item{object}{color_class object.}

\item{...}{fit2 - a second color_class object, rgb - color defined through rgb, hsv - color defined through rgb, rope - region of practical equivalence, par - components of comparison, a subset of (r, g, b, h, s, v).}
}
\description{
\code{compare} prints difference in colors between two fits, or a fit and a color.
}
