% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/color_class.R
\name{compare_means,color_class-method}
\alias{compare_means,color_class-method}
\alias{compare_means_color}
\title{compare_means}
\usage{
\S4method{compare_means}{color_class}(object, ...)
}
\arguments{
\item{object}{color_class object.}

\item{...}{fit2 - a second color_class object, rgb - color defined through rgb, hsv - color defined through rgb, rope - region of practical equivalence, pars - components of comparison, a subset of (r, g, b, h, s, v).}
}
\value{
Comparison results or an error if something went wrong.
}
\description{
\code{compare_means} prints difference in colors between two fits or a fit and a color.
}
\examples{
# to use the function you first have to prepare the data and fit the model
# see class documentation for an example of the whole process
# along with an example of how to use this function
?color_class

}
