% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/color_class.R
\docType{methods}
\name{get_parameters,color_class-method}
\alias{get_parameters,color_class-method}
\alias{get_parameters_color_class}
\title{get_parameters}
\usage{
\S4method{get_parameters}{color_class}(object)
}
\arguments{
\item{object}{color_class object.}
}
\description{
\code{get_parameters} returns a dataframe with values of fitted parameters.
}
