% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/color_class.R
\docType{methods}
\name{get_samples,color_class-method}
\alias{get_samples,color_class-method}
\alias{get_samples_color_class}
\title{get_samples}
\usage{
\S4method{get_samples}{color_class}(object)
}
\arguments{
\item{object}{color_class object.}
}
\description{
\code{get_samples} returns a dataframe with values of fitted parameters.
}
