% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/color_class.R
\docType{methods}
\name{plot_difference,color_class-method}
\alias{plot_difference,color_class-method}
\alias{plot_difference_color}
\title{plot_difference}
\usage{
\S4method{plot_difference}{color_class}(object, ...)
}
\arguments{
\item{object}{color_class object.}

\item{...}{fit2 - a second color_class object, rgb - color defined through rgb, hsv - color defined through rgb, rope - region of practical equivalence, bins - number of bins in the histogram, par - components of comparison - a subset of (r, g, b, h, s, v).}
}
\description{
\code{plot_difference} a visualization of the difference between two fits
}
