% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/color_class.R
\docType{methods}
\name{plot_distributions,color_class-method}
\alias{plot_distributions,color_class-method}
\alias{plot_distributions_color}
\title{plot_distributions}
\usage{
\S4method{plot_distributions}{color_class}(object, ...)
}
\arguments{
\item{object}{color_class object.}

\item{...}{fit2 - a second color_class object, rgb - color defined through rgb, hsv - color defined through rgb, par - components of comparison, a subset of (r, g, b, h, s, v).}
}
\description{
\code{plot_distributions} a visualization of the fitted distributions or constant colors.
}
