% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/color_class.R
\docType{methods}
\name{plot_distributions_difference,color_class-method}
\alias{plot_distributions_difference,color_class-method}
\alias{plot_distributions_difference_color}
\title{plot_distributions_difference}
\usage{
\S4method{plot_distributions_difference}{color_class}(object, ...)
}
\arguments{
\item{object}{color_class object.}

\item{...}{fit2 - a second linear_class object, rope - region of practical equivalence, bins - number of bins in the histogram.}
}
\description{
\code{plot_distributions_difference} a visualization of the difference between the distribution of the first group and the second group.
}
