% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/color_class.R
\docType{methods}
\name{plot_distributions_hsv}
\alias{plot_distributions_hsv}
\alias{plot_distributions_hsv,color_class-method}
\alias{plot_distributions_hsv_color}
\title{plot_distributions_hsv}
\usage{
plot_distributions_hsv(object, ...)

\S4method{plot_distributions_hsv}{color_class}(object, ...)
}
\arguments{
\item{object}{color_class object.}

\item{...}{fit2 - a second color_class object, points - points to plot defined through rgb or hsv, lines - lines to plot defined through rgb or hsv, hsv - are points and lines defined in hsv format (default = FALSE).}
}
\description{
\code{plot_distributions_hsv} a visualization of distributions of one or two fits thorough a circular visualization of hsv color components. You can also compare fit means with colors defined through rgb or hsv components (as points or as lines on the visualization).
}
