% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/color_class.R
\docType{methods}
\name{plot_fit,color_class-method}
\alias{plot_fit,color_class-method}
\alias{plot_fit_color}
\title{plot_fit}
\usage{
\S4method{plot_fit}{color_class}(object, ...)
}
\arguments{
\item{object}{color_class object.}

\item{...}{par - components of comparison, a subset of (r, g, b, h, s, v).}
}
\description{
\code{plot_fit} plots fitted model against the data. Use this function to explore the quality of your fit. You can compare fit with underlying data only through chosen color components (r, g, b, h, s, v).
}
