% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/color_class.R
\docType{methods}
\name{plot_fit_hsv}
\alias{plot_fit_hsv}
\alias{plot_fit_hsv,color_class-method}
\alias{plot_fit_hsv_color}
\title{plot_fit_hsv}
\usage{
plot_fit_hsv(object)

\S4method{plot_fit_hsv}{color_class}(object)
}
\arguments{
\item{object}{color_class object.}
}
\description{
\code{plot_fit_hsv} plots fitted model against the data. Use this function to explore the quality of your fit thorough a circular visualization of hsv color components.
}
