% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/color_class.R
\docType{methods}
\name{plot_hsvfit}
\alias{plot_hsvfit}
\alias{plot_hsvfit,color_class-method}
\alias{plot_hsvfit_color}
\title{plot_hsvfit}
\usage{
plot_hsvfit(object)

\S4method{plot_hsvfit}{color_class}(object)
}
\arguments{
\item{object}{color_class object.}

\item{object}{color_class object.}
}
\description{
\code{plot_hsvfit} plots fitted model against the data. Use this function to explore the quality of your fit thorough a circular visualization of hsv color components.

\code{plot_hsvfit} plots fitted model against the data. Use this function to explore the quality of your fit thorough a circular visualization of hsv color components.
}
