% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/color_class.R
\docType{methods}
\name{plot_samples,color_class-method}
\alias{plot_samples,color_class-method}
\alias{plot_samples_color}
\title{plot_samples}
\usage{
\S4method{plot_samples}{color_class}(object, ...)
}
\arguments{
\item{object}{color_class object.}

\item{...}{fit2 - a second color_class object, rgb - color defined through rgb, hsv - color defined through rgb, pars - components of comparison, a subset of (r, g, b, h, s, v).}
}
\description{
\code{plot_samples} lots density of the samples, the first and the second group samples or a constant values in case second group is defined as rgb or hsv color..
}
