% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/color_class.R
\docType{methods}
\name{plot_trace,color_class-method}
\alias{plot_trace,color_class-method}
\alias{plot_trace_color}
\title{plot_trace}
\usage{
\S4method{plot_trace}{color_class}(object)
}
\arguments{
\item{object}{color_class object.}
}
\description{
\code{plot_trace} traceplot for main fitted model parameters.
}
