% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/success_rate_class.R
\docType{methods}
\name{get_parameters,success_rate_class-method}
\alias{get_parameters,success_rate_class-method}
\title{get_parameters}
\usage{
\S4method{get_parameters}{success_rate_class}(object)
}
\arguments{
\item{object}{success_rate_class object.}
}
\description{
\code{get_parameters} returns a dataframe with values of fitted parameters.
}
