% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linear_class.R
\docType{class}
\name{linear_class-class}
\alias{linear_class-class}
\alias{linear_class}
\title{linear_class}
\description{
An S4 class for storing results of normal linear model.

\strong{Functions}

summary(`linear_class`): prints summary of the fit.

print(`linear_class`): prints a more detailed summary of the fit

show(`linear_class`): prints a more detailed summary of the fit.

get_parameters(`linear_class`): returns a dataframe with values of fitted parameters.

get_subject_parameters(`linear_class`): returns a dataframe with values of fitted parameters for each subject in the hierarchical model.

compare_means(`linear_class`, fit2=`linear_class`): prints difference in slope and intercept between two groups. You can also provide the rope parameter.

plot_means_difference(`linear_class`, fit2=`linear_class`): a visualization of the difference between two groups. You can also provide the rope and bins (number of bins in the histogram) parameters.

plot_means(`linear_class`): plots density of means.

plot_means(`linear_class`, fit2=`linear_class`): plots density for the first and the second group means.

compare_distributions(`linear_class`, fit2=`linear_class`): draws samples from distribution of the first group and compares them against samples drawn from the distribution of the second group.

plot_distributions(`linear_class`): a visualization of the fitted distribution.

plot_distributions(`linear_class`, fit2=`linear_class`): a visualization of two fitted distribution.

plot_distributions_difference(`linear_class`, fit2=`linear_class`): a visualization of the difference between the distribution of the first group and the second group. You can also provide the rope and bins (number of bins in the histogram) parameters.

plot_fit(`linear_class`): plots fitted model against the data. Use this function to explore the quality of your fit. Fit will be plotted on the group level.

plot_fit(`linear_class`, subjects='boolean'): plots fitted model against the data. Use this function to explore the quality of your fit. You can plot on the group level (subjects=FALSE) or on the subjects level (subjects=TRUE).

plot_trace(`linear_class`): traceplot for main fitted model parameters.
}
\section{Slots}{

\describe{
\item{\code{extract}}{Extract from Stan fit.}

\item{\code{fit}}{Stan fit.}

\item{\code{data}}{Raw data for the tested group.}
}}

