% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linear_class.R
\docType{methods}
\name{traceplot,linear_class-method}
\alias{traceplot,linear_class-method}
\alias{compare,ANY-method}
\title{compare}
\usage{
\S4method{traceplot}{linear_class}(object)
}
\description{
\code{compare} prints difference/equality of the first group against the second group, against a mean value, or against a normal distribution with a defined mean value and variance.

\code{plot_difference} a visualization of the difference of the first group against the second group, against a mean value, or against a normal distribution with a defined mean value and variance.

\code{plot_comparison} plots density for the first group and density for the second group, or a mean value in case second group is defined as a normal distribution or as a constant.

\code{compare_distributions} draws samples from distribution of the first group and compares them against samples drawn from the distribution of the second group, against a mean value, or against samples from a normal distribution with a defined mean value and variance.

\code{plot_distributions} visualizes distributions underlying tested groups.

\code{plot_distributions_difference} a visualization of the difference between the distribution of the first group and the distribution or a constant value for the second group.

\code{plot_fit} plots fitted model against the data. Use this function to explore the quality of your fit.

\code{traceplot} traceplot for main fitted model parameters.
}
