% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linear_class.R
\docType{methods}
\name{plot_fit,linear_class-method}
\alias{plot_fit,linear_class-method}
\alias{compare_distributions,ANY-method}
\title{compare_distributions}
\usage{
\S4method{plot_fit}{linear_class}(object)
}
\description{
\code{compare_distributions} TODO

\code{plot_distributions} TODO

\code{plot_distributions_difference} TODO

\code{plot_fit} plots fitted model against the data. Use this function to explore the quality of your fit.
}
