% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linear_class.R
\docType{methods}
\name{compare_means,linear_class-method}
\alias{compare_means,linear_class-method}
\alias{compare_meanslinear}
\title{compare_means}
\usage{
\S4method{compare_means}{linear_class}(object, ...)
}
\arguments{
\item{object}{linear_class object.}

\item{...}{fit2 - a second linear_class object, rope_intercept and rope_slope - regions of practical equivalence.}
}
\description{
\code{compare_means} prints difference in intercept and slope between two groups.
}
