% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linear_class.R
\docType{methods}
\name{compare_samples,linear_class-method}
\alias{compare_samples,linear_class-method}
\alias{compare_sampleslinear}
\title{compare_samples}
\usage{
\S4method{compare_samples}{linear_class}(object, ...)
}
\arguments{
\item{object}{linear_class object.}

\item{...}{fit2 - a second linear_class object, rope_intercept and rope_slope - regions of practical equivalence.}
}
\description{
\code{compare_samples} prints difference in intercept and slope between two groups.
}
