% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linear_class.R
\docType{methods}
\name{get_subject_parameters,linear_class-method}
\alias{get_subject_parameters,linear_class-method}
\alias{get_subject_parameters_linear_class}
\title{get_subject_parameters}
\usage{
\S4method{get_subject_parameters}{linear_class}(object)
}
\arguments{
\item{object}{linear_class object.}
}
\description{
\code{get_subject_parameters} returns a dataframe with values of fitted parameters for each subject in the hierarchical model.
}
