% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linear_class.R
\docType{methods}
\name{get_subject_samples,linear_class-method}
\alias{get_subject_samples,linear_class-method}
\alias{get_subject_samples_linear_class}
\title{get_subject_samples}
\usage{
\S4method{get_subject_samples}{linear_class}(object)
}
\arguments{
\item{object}{linear_class object.}
}
\description{
\code{get_subject_samples} returns a dataframe with values of fitted parameters for each subject in the hierarchical model.
}
