% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linear_class.R
\docType{methods}
\name{plot_difference,linear_class-method}
\alias{plot_difference,linear_class-method}
\alias{plot_difference,ANY-method}
\title{plot_difference}
\usage{
\S4method{plot_difference}{linear_class}(object, ...)
}
\description{
\code{plot_difference} TODO
}
