% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linear_class.R
\docType{methods}
\name{plot_fit,linear_class-method}
\alias{plot_fit,linear_class-method}
\alias{plot_fit,-method,ANY-method}
\title{plot_fit}
\usage{
\S4method{plot_fit}{linear_class}(object)
}
\arguments{
\item{object}{linear_class object.}
}
\description{
\code{plot_fit} plots fitted model against the data. Use this function to explore the quality of your fit.
}
