% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linear_class.R
\docType{methods}
\name{plot_means,linear_class-method}
\alias{plot_means,linear_class-method}
\alias{plot_means_linear}
\title{plot_means}
\usage{
\S4method{plot_means}{linear_class}(object, ...)
}
\arguments{
\item{object}{linear_class object.}

\item{...}{fit2 - a second linear_class object.}
}
\description{
\code{plot_means} plots means or the first and the second group means.
}
