% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linear_class.R
\docType{methods}
\name{plot_means_difference,linear_class-method}
\alias{plot_means_difference,linear_class-method}
\alias{plot_means_difference_linear}
\title{plot_means_difference}
\usage{
\S4method{plot_means_difference}{linear_class}(object, ...)
}
\arguments{
\item{object}{linear_class object.}

\item{...}{fit2 - a second linear_class object, par - specific parameter of comparison (slope or intercept), rope_intercept and rope_slope - regions of practical equivalence, bins - number of bins in the histogram.}
}
\description{
\code{plot_means_difference} plots difference between two groups.
}
