% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linear_class.R
\docType{methods}
\name{plot_samples,linear_class-method}
\alias{plot_samples,linear_class-method}
\alias{plot_samples_linear}
\title{plot_samples}
\usage{
\S4method{plot_samples}{linear_class}(object, ...)
}
\arguments{
\item{object}{linear_class object.}

\item{...}{fit2 - a second linear_class object.}
}
\description{
\code{plot_samples} plots samples or the first and the second group samples.
}
