% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linear_class.R
\docType{methods}
\name{plot_trace,linear_class-method}
\alias{plot_trace,linear_class-method}
\alias{plot_trace,linear_class-method,ANY-method}
\title{plot_trace}
\usage{
\S4method{plot_trace}{linear_class}(object)
}
\arguments{
\item{object}{linear_class object.}
}
\description{
\code{plot_trace} traceplot for main fitted model parameters.
}
