% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linear_class.R
\docType{methods}
\name{traceplot,linear_class-method}
\alias{traceplot,linear_class-method}
\alias{traceplot,ANY-method}
\title{traceplot}
\usage{
\S4method{traceplot}{linear_class}(object)
}
\description{
\code{traceplot} traceplot for main fitted model parameters.
}
