% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normal_stan.R
\name{normal_stan}
\alias{normal_stan}
\title{Bayesian normal model}
\usage{
normal_stan(y, yMu = 0, ySd = 1, uL = 0, uH = 2, iter = 1000)
}
\arguments{
\item{y}{Numeric vector of values.}

\item{...}{Arguments passed to `rstan::sampling` (e.g. iter, chains).}
}
\value{
An object of class `stanfit` returned by `rstan::sampling`
}
\description{
Bayesian normal model
}
