% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/color_class.R
\name{plot,color_class,missing-method}
\alias{plot,color_class,missing-method}
\title{plot}
\usage{
\S4method{plot}{color_class,missing}(x, y, ...)
}
\arguments{
\item{x}{color_class object.}

\item{y}{empty dummy variable, ignore this.}

\item{...}{pars - components of comparison, a subset of (r, g, b, h, s, v).}
}
\description{
\code{plot} plots fitted model against the data. Use this function to explore the quality of your fit. You can compare fit with underlying data only through chosen color components (r, g, b, h, s, v).
}
\examples{
# to use the function you first have to prepare the data and fit the model
# see class documentation for an example of the whole process
# along with an example of how to use this function
?color_class

}
