% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reaction_time_class.R
\name{plot,reaction_time_class,missing-method}
\alias{plot,reaction_time_class,missing-method}
\title{plot}
\usage{
\S4method{plot}{reaction_time_class,missing}(x, y, ...)
}
\arguments{
\item{x}{reaction_time_class object.}

\item{y}{empty dummy variable, ignore this.}

\item{...}{subjects - plot fits on a subject level (default = TRUE).}
}
\description{
\code{plot} plots fitted model against the data. Use this function to explore the quality of your fit. You can plot on the subjects level (subjects=TRUE) or on the group level (subjects=FALSE).
}
\examples{
# to use the function you first have to prepare the data and fit the model
# see class documentation for an example of the whole process
# along with an example of how to use this function
?reaction_time_class

}
