% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/success_rate_class.R
\docType{methods}
\name{plot,success_rate_class,missing-method}
\alias{plot,success_rate_class,missing-method}
\title{plot}
\usage{
\S4method{plot}{success_rate_class,missing}(x, y, ...)
}
\arguments{
\item{...}{subjects - plot fits on a subject level (default = TRUE).}

\item{object}{success_rate_class object.}
}
\description{
\code{plot} plots fitted model against the data. Use this function to explore the quality of your fit. You can plot on the subjects level (subjects=TRUE) or on the group level (subjects=FALSE).
}
\examples{
# to use the function you first have to prepare the data and fit the model
# see class documentation for an example of the whole process
# along with an example of how to use this function
?success_rate_class

}
