% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ttest_class.R
\name{plot,ttest_class,missing-method}
\alias{plot,ttest_class,missing-method}
\title{plot}
\usage{
\S4method{plot}{ttest_class,missing}(x)
}
\arguments{
\item{x}{ttest_class object.}
}
\description{
\code{plot} plots fitted model against the data. Use this function to explore the quality of your fit.
}
\examples{
# to use the function you first have to prepare the data and fit the model
# see class documentation for an example of the whole process
# along with an example of how to use this function
?ttest_class

}
