% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shared_functions.R
\name{prepare_rope}
\alias{prepare_rope}
\title{prepare_rope}
\usage{
prepare_rope(rope)
}
\arguments{
\item{rope}{Rope interval parameter (single number, or an interval).}
}
\value{
Rope as an interval.
}
\description{
A shared function for all classes that casts rope interval into a suitable format.
}
