% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reaction_time_class.R
\docType{methods}
\name{compare,reaction_time_class-method}
\alias{compare,reaction_time_class-method}
\alias{compare,ANY-method}
\title{compare}
\usage{
\S4method{compare}{reaction_time_class}(object, ...)
}
\description{
\code{compare} prints difference in reaction times between two groups.
}
