% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reaction_time_class.R
\docType{methods}
\name{compare_distributions,reaction_time_class-method}
\alias{compare_distributions,reaction_time_class-method}
\alias{compare_distributions,ANY-method}
\title{compare_distributions}
\usage{
\S4method{compare_distributions}{reaction_time_class}(object, ...)
}
\description{
\code{compare_distributions} draws samples from distribution of the first group and compares them against samples drawn from the distribution of the second group.
}
