% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reaction_time_class.R
\docType{methods}
\name{compare_samples,reaction_time_class-method}
\alias{compare_samples,reaction_time_class-method}
\alias{compare_samples_reaction_time}
\title{compare_samples}
\usage{
\S4method{compare_samples}{reaction_time_class}(object, ...)
}
\arguments{
\item{object}{reaction_time_class object.}

\item{...}{fit2 - a second reaction_time_class object, rope - region of practical equivalence, par - specific parameter of comparison - mu or lambda.}
}
\description{
\code{compare_samples} prints difference in reaction times between two groups.
}
