% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reaction_time_class.R
\docType{methods}
\name{get_samples,reaction_time_class-method}
\alias{get_samples,reaction_time_class-method}
\alias{get_samples_reaction_time}
\title{get_samples}
\usage{
\S4method{get_samples}{reaction_time_class}(object)
}
\arguments{
\item{object}{reaction_time_class object.}
}
\description{
\code{get_samples} returns a dataframe with values of fitted parameters.
}
