% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reaction_time_class.R
\docType{methods}
\name{get_subject_parameters,reaction_time_class-method}
\alias{get_subject_parameters,reaction_time_class-method}
\alias{get_subject_parameters_reaction_time}
\title{get_subject_parameters}
\usage{
\S4method{get_subject_parameters}{reaction_time_class}(object)
}
\arguments{
\item{object}{reaction_time_class object.}
}
\description{
\code{get_subject_parameters} returns a dataframe with values of fitted parameters for each subject in the hierarchical model.
}
