% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reaction_time_class.R
\docType{methods}
\name{plot_comparison,reaction_time_class-method}
\alias{plot_comparison,reaction_time_class-method}
\alias{plot_comparison,ANY-method}
\title{plot_comparison}
\usage{
\S4method{plot_comparison}{reaction_time_class}(object, ...)
}
\description{
\code{plot_comparison} plots density for the first and the second group.
}
