% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reaction_time_class.R
\docType{methods}
\name{plot_difference,reaction_time_class-method}
\alias{plot_difference,reaction_time_class-method}
\alias{plot_difference,ANY-method}
\title{plot_difference}
\usage{
\S4method{plot_difference}{reaction_time_class}(object, ...)
}
\description{
\code{plot_difference} a visualization of the difference between two groups.
}
