% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reaction_time_class.R
\docType{methods}
\name{plot_distributions,reaction_time_class-method}
\alias{plot_distributions,reaction_time_class-method}
\alias{plot_distributions_reaction_time}
\title{plot_distributions}
\usage{
\S4method{plot_distributions}{reaction_time_class}(object, ...)
}
\arguments{
\item{object}{reaction_time_class object.}

\item{...}{fit2 - a second reaction_time_class object.}
}
\description{
\code{plot_distributions} a visualization of the fitted distribution, for one or two fits.
}
