% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reaction_time_class.R
\docType{methods}
\name{plot_distributions_difference,reaction_time_class-method}
\alias{plot_distributions_difference,reaction_time_class-method}
\alias{plot_distributions_difference_reaction_time}
\title{plot_distributions_difference}
\usage{
\S4method{plot_distributions_difference}{reaction_time_class}(object, ...)
}
\arguments{
\item{object}{reaction_time_class object.}

\item{...}{fit2 - a second reaction_time_class object, fits - a list of reaction_time_class objects, rope - region of practical equivalence, bins - number of bins in the histogram.}
}
\description{
\code{plot_distributions_difference} a visualization of the difference between the distribution of the first group and the second group or between multiple groups.
}
