% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reaction_time_class.R
\docType{methods}
\name{plot_fit,reaction_time_class-method}
\alias{plot_fit,reaction_time_class-method}
\alias{plot_fit,ANY-method}
\title{plot_fit}
\usage{
\S4method{plot_fit}{reaction_time_class}(object)
}
\description{
\code{plot_fit} plots fitted model against the data. Use this function to explore the quality of your fit.
}
