% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reaction_time_class.R
\docType{methods}
\name{plot_means,reaction_time_class-method}
\alias{plot_means,reaction_time_class-method}
\alias{plot_means_reaction_time}
\title{plot_means}
\usage{
\S4method{plot_means}{reaction_time_class}(object, ...)
}
\arguments{
\item{object}{reaction_time_class object.}

\item{...}{fit2 - a second reaction_time_class object, fits - a list of reaction_time_class objects, par - specific parameter of comparison - mu or lambda.}
}
\description{
\code{plot_means} plots density of means for one, two or multiple groups.
}
