% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reaction_time_class.R
\docType{methods}
\name{plot_samples,reaction_time_class-method}
\alias{plot_samples,reaction_time_class-method}
\alias{plot_samples,ANY-method}
\title{plot_samples}
\usage{
\S4method{plot_samples}{reaction_time_class}(object, ...)
}
\description{
\code{plot_samples} plots density for the first, or the first and the second group.
}
