% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reaction_time_class.R
\docType{methods}
\name{plot_samples_difference,reaction_time_class-method}
\alias{plot_samples_difference,reaction_time_class-method}
\alias{plot_samples_difference_reaction_time}
\title{plot_samples_difference}
\usage{
\S4method{plot_samples_difference}{reaction_time_class}(object, ...)
}
\arguments{
\item{object}{reaction_time_class object.}

\item{...}{fit2 - a second reaction_time_class object, fits - a list of reaction_time_class objects, rope - region of practical equivalence, bins - number of bins in the histogram, par - specific parameter of comparison - mu or lambda.}
}
\description{
\code{plot_samples_difference} a visualization of the difference between two groups or multiple groups.
}
