% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reaction_time_class.R
\docType{methods}
\name{plot_trace,reaction_time_class-method}
\alias{plot_trace,reaction_time_class-method}
\alias{plot_trace_reaction_time}
\title{plot_trace}
\usage{
\S4method{plot_trace}{reaction_time_class}(object)
}
\arguments{
\item{object}{reaction_time_class object.}
}
\description{
\code{plot_trace} traceplot for main fitted model parameters.
}
