% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reaction_time_class.R
\docType{methods}
\name{traceplot,reaction_time_class-method}
\alias{traceplot,reaction_time_class-method}
\alias{traceplot,ANY-method}
\title{traceplot}
\usage{
\S4method{traceplot}{reaction_time_class}(object)
}
\description{
\code{traceplot} traceplot for main fitted model parameters.
}
