% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reaction_times_results_class.R
\docType{class}
\name{reaction_times_results-class}
\alias{reaction_times_results-class}
\alias{reaction_times_results}
\title{An S4 class for storing results of reaction time Bayesian model.}
\description{
An S4 class for storing results of reaction time Bayesian model.
}
\section{Slots}{

\describe{
\item{\code{extract}}{Extract from Stan fit.}

\item{\code{fit}}{Stan fit.}

\item{\code{data}}{Data on which the fit is based.}
}}

\examples{
summary(`reaction_times_results`): prints summary od the fit.

compare(`reaction_times_results`, `reaction_times_results`): prints difference in reaction times between two groups.

plot_difference(`reaction_times_results`, `reaction_times_results`): a visualization of the difference between two groups.

plot_comparison(`reaction_times_results`, `reaction_times_results`): plots density for the first and the second group.

compare_distributions(`reaction_times_results`, `reaction_times_results`): draws samples from distribution of the first group and compares them against samples drawn from the distribution of the second group.

plot_distributions(`reaction_times_results`, `reaction_times_results`): a visualization of the distribution for the first group and the second group.

plot_distributions_difference(`reaction_times_results`, `reaction_times_results`): a visualization of the difference between the distribution of the first group and the second group.

plot_fit(`reaction_times_results`): plots fitted model against the data. Use this function to explore the quality of your fit.

traceplot(`reaction_times_results`): traceplot for main fitted model parameters.

}
