% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reaction_times_results_class.R
\docType{methods}
\name{compare,reaction_times_results-method}
\alias{compare,reaction_times_results-method}
\alias{compare,ANY-method}
\title{compare}
\usage{
\S4method{compare}{reaction_times_results}(object, ...)
}
\description{
\code{compare} prints difference in reaction times between two groups.
}
