% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reaction_times_results_class.R,
%   R/successes_results_class.R
\docType{methods}
\name{plot_distributions,reaction_times_results-method}
\alias{plot_distributions,reaction_times_results-method}
\alias{plot_distributions,ANY-method}
\alias{plot_distributions,successes_results-method}
\title{plot_distributions}
\usage{
\S4method{plot_distributions}{reaction_times_results}(object, ...)

\S4method{plot_distributions}{successes_results}(object, ...)
}
\description{
\code{plot_distributions} a visualization of the distribution for the first group and the second group.

\code{plot_distributions} a visualization of the distribution for the first group and the second group.
}
