% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shared_functions.R
\name{shared_difference}
\alias{shared_difference}
\title{Shared helper functions used in several models.}
\usage{
shared_difference(y1, y2, rope = NULL)
}
\arguments{
\item{y1}{Numeric vector of values for the first group.}

\item{y2}{Numeric vector of values for the second group.}

\item{rope}{Optional parameter for the rope interval.}
}
\value{
Prints the difference and 95% confidence interval into the console.
}
\description{
Shared helper functions used in several models.
}
