% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/color_class.R
\docType{methods}
\name{show,color_class-method}
\alias{show,color_class-method}
\title{show}
\usage{
\S4method{show}{color_class}(object)
}
\arguments{
\item{object}{color_class object.}
}
\description{
\code{show} prints a more detailed summary of the Bayesian color fit.
}
