% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linear_class.R
\docType{methods}
\name{show,linear_class-method}
\alias{show,linear_class-method}
\title{show}
\usage{
\S4method{show}{linear_class}(object)
}
\arguments{
\item{object}{linear_class object.}
}
\description{
\code{show} prints a more detailed summary of the Bayesian linear model fit.
}
