% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reaction_time_class.R
\docType{methods}
\name{show,reaction_time_class-method}
\alias{show,reaction_time_class-method}
\title{show}
\usage{
\S4method{show}{reaction_time_class}(object)
}
\arguments{
\item{object}{reaction_time_class object.}
}
\description{
\code{show} prints a more detailed summary of the Bayesian success rate fit.
}
