% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/success_rate_class.R
\docType{methods}
\name{show,success_rate_class-method}
\alias{show,success_rate_class-method}
\title{show}
\usage{
\S4method{show}{success_rate_class}(object)
}
\arguments{
\item{object}{success_rate_class object.}
}
\description{
\code{show} prints a more detailed summary of the Bayesian success rate fit.
}
