% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ttest_class.R
\docType{methods}
\name{show,ttest_class-method}
\alias{show,ttest_class-method}
\title{show}
\usage{
\S4method{show}{ttest_class}(object)
}
\arguments{
\item{object}{ttest_class object.}
}
\description{
\code{show} prints a more detailed summary of the Bayesian ttest fit.
}
