% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/success_rate_class.R
\docType{class}
\name{success_rate_class-class}
\alias{success_rate_class-class}
\alias{success_rate_class}
\title{success_rate_class}
\description{
An S4 class for storing results of successes (true/false) Bayesian model.

\strong{Functions}

summary(`success_rate_class`): prints summary of the fit.

print(`success_rate_class`): prints a more detailed summary of the fit

show(`success_rate_class`): prints a more detailed summary of the fit.

get_parameters(`success_rate_class`): returns a dataframe with values of fitted parameters.

get_subject_parameters(`success_rate_class`): returns a dataframe with values of fitted parameters for each subject in the hierarchical model.

compare_means(`success_rate_class`, fit2=`success_rate_class`): returns difference in successfulness between two groups. You can also provide the rope parameter.

compare_means(`success_rate_class`, fits=`list`): returns difference in successfulness between multiple groups. You can also provide the rope parameter.

plot_means_difference(`success_rate_class`, fit2=`success_rate_class`): a visualization of the difference between two groups. You can also provide the rope and bins (number of bins in the histogram) parameters.

plot_means_difference(`success_rate_class`, fits=`list`): a visualization of the difference between multiple groups. You can also provide the rope and bins (number of bins in the histogram) parameters.

plot_means(`success_rate_class`): plots density for the first group means.

plot_means(`success_rate_class`, fit2=`success_rate_class`): plots density for the first and the second group means.

plot_means(`success_rate_class`, fits=`list`): plots density for multiple

compare_distributions(`success_rate_class`, fit2=`success_rate_class`): draws samples from distribution of the first group and compares them against samples drawn from the distribution of the second group. You can also provide the rope parameter.

compare_distributions(`success_rate_class`, fits=`list`): draws and compares samples from distributions of multiple groups. You can also provide the rope parameter.

plot_distributions(`success_rate_class`): a visualization of the fitted distribution.

plot_distributions(`success_rate_class`, fit2=`success_rate_class`): a visualization of the distribution for two fits.

plot_distributions(`success_rate_class`, fits=`list`): a visualization of the distribution for multiple fits.

plot_distributions_difference(`success_rate_class`, fit2=`success_rate_class`): a visualization of the difference between the distribution of the first group and the second group. You can also provide the rope and bins (number of bins in the histogram) parameters.

plot_distributions_difference(`success_rate_class`, fits=`list`): a visualization of the difference between the distributions of multiple groups. You can also provide the rope and bins (number of bins in the histogram) parameters.

plot_fit(`success_rate_class`): plots fitted model against the data. Use this function to explore the quality of your fit. Fit will be plotted on the group level.

plot_fit(`success_rate_class`, subjects='boolean'): plots fitted model against the data. Use this function to explore the quality of your fit. You can plot on the group level (subjects=FALSE) or on the subjects level (subjects=TRUE).

plot_trace(`success_rate_class`): traceplot for main fitted model parameters.
}
\section{Slots}{

\describe{
\item{\code{extract}}{Extract from Stan fit.}

\item{\code{fit}}{Stan fit.}

\item{\code{data}}{Data on which the fit is based.}
}}

