% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/success_rate_class.R
\docType{methods}
\name{compare,success_rate_class-method}
\alias{compare,success_rate_class-method}
\alias{compare,ANY-method}
\title{compare}
\usage{
\S4method{compare}{success_rate_class}(object, ...)
}
\description{
\code{compare} prints difference in successfulness between two groups.
}
