% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/success_rate_class.R
\docType{methods}
\name{compare_means,success_rate_class-method}
\alias{compare_means,success_rate_class-method}
\alias{compare_means_success_rate}
\title{compare_means}
\usage{
\S4method{compare_means}{success_rate_class}(object, ...)
}
\arguments{
\item{object}{success_rate_class object.}

\item{...}{fit2 - a second success_rate_class object, fits - a list of success_rate_class objects, rope - region of practical equivalence.}
}
\description{
\code{compare_means} prints difference in success rate between two groups or multiple groups.
}
