% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/success_rate_class.R
\docType{methods}
\name{get_samples,success_rate_class-method}
\alias{get_samples,success_rate_class-method}
\alias{get_samples_success_rate_class}
\title{get_samples}
\usage{
\S4method{get_samples}{success_rate_class}(object)
}
\arguments{
\item{object}{success_rate_class object.}
}
\description{
\code{get_samples} returns a dataframe with values of fitted parameters.
}
