% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/success_rate_class.R
\docType{methods}
\name{plot_comparison,success_rate_class-method}
\alias{plot_comparison,success_rate_class-method}
\alias{plot_comparison,ANY-method}
\title{plot_comparison}
\usage{
\S4method{plot_comparison}{success_rate_class}(object, ...)
}
\description{
\code{plot_comparison} plots density for the first and the second group.
}
