% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/success_rate_class.R
\docType{methods}
\name{plot_means,success_rate_class-method}
\alias{plot_means,success_rate_class-method}
\alias{plot_means_success_rate}
\title{plot_means}
\usage{
\S4method{plot_means}{success_rate_class}(object, ...)
}
\arguments{
\item{object}{success_rate_class object.}

\item{...}{fit2 - a second success_rate_class object, fits - a list of success_rate_class objects.}
}
\description{
\code{plot_means} plots density of means for one, two or multiple groups.
}
