% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/success_rate_class.R
\docType{methods}
\name{plot_means_difference,success_rate_class-method}
\alias{plot_means_difference,success_rate_class-method}
\alias{plot_means_difference_success_rate}
\title{plot_means_difference}
\usage{
\S4method{plot_means_difference}{success_rate_class}(object, ...)
}
\arguments{
\item{object}{success_rate_class object.}

\item{...}{fit2 - a second success_rate_class object, fits - a list of success_rate_class objects, rope - region of practical equivalence, bins - number of bins in the histogram.}
}
\description{
\code{plot_means_difference} a visualization of the difference between two groups or multiple groups.
}
