% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/success_rate_class.R
\docType{methods}
\name{plot_trace,success_rate_class-method}
\alias{plot_trace,success_rate_class-method}
\alias{plot_trace_success_rate}
\title{plot_trace}
\usage{
\S4method{plot_trace}{success_rate_class}(object)
}
\arguments{
\item{object}{success_rate_class object.}
}
\description{
\code{plot_trace} traceplot for main fitted model parameters.
}
