% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/success_rate_class.R
\docType{methods}
\name{traceplot,success_rate_class-method}
\alias{traceplot,success_rate_class-method}
\alias{traceplot,ANY-method}
\title{traceplot}
\usage{
\S4method{traceplot}{success_rate_class}(object)
}
\description{
\code{traceplot} traceplot for main fitted model parameters.
}
