% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/successes_results_class.R
\docType{class}
\name{successes_results-class}
\alias{successes_results-class}
\alias{successes_results}
\title{An S4 class for storing results of successes (true/false) Bayesian model.}
\description{
An S4 class for storing results of successes (true/false) Bayesian model.
}
\section{Slots}{

\describe{
\item{\code{extract}}{Extract from Stan fit.}

\item{\code{fit}}{Stan fit.}

\item{\code{data}}{Data on which the fit is based.}
}}

\examples{
summary(`successes_results`): prints summary od the fit.

compare(`successes_results`, `successes_results`): prints difference in successfulness of two groups.

plot_difference(`successes_results`, `successes_results`): a visualization of the difference between two groups.

plot_comparison(`successes_results`, `successes_results`): plots density for the first and the second group.

compare_distributions(`successes_results`, `successes_results`): draws samples from distribution of the first group and compares them against samples drawn from the distribution of the second group.

plot_distributions(`successes_results`, `successes_results`): a visualization of the distribution for the first group and the second group.

plot_distributions_difference(`successes_results`, `successes_results`): a visualization of the difference between the distribution of the first group and the second group.

plot_fit(`successes_results`): plots fitted model against the data. Use this function to explore the quality of your fit.

traceplot(`successes_results`): traceplot for main fitted model parameters.

}
