% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/successes_results_class.R
\docType{methods}
\name{compare,successes_results-method}
\alias{compare,successes_results-method}
\alias{compare,ANY-method}
\title{compare}
\usage{
\S4method{compare}{successes_results}(object, ...)
}
\description{
\code{compare} prints difference in successfulness between two groups.
}
