% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/successes_results_class.R
\docType{methods}
\name{compare_distributions,successes_results-method}
\alias{compare_distributions,successes_results-method}
\alias{compare_distributions,ANY-method}
\title{compare_distributions}
\usage{
\S4method{compare_distributions}{successes_results}(object, ...)
}
\description{
\code{compare_distributions} draws samples from distribution of the first group and compares them against samples drawn from the distribution of the second group.
}
