% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/successes_results_class.R
\docType{methods}
\name{plot_difference,successes_results-method}
\alias{plot_difference,successes_results-method}
\alias{plot_difference,ANY-method}
\title{plot_difference}
\usage{
\S4method{plot_difference}{successes_results}(object, ..., bins = 30)
}
\description{
\code{plot_difference} a visualization of the difference between two groups.
}
