% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/successes_results_class.R
\docType{methods}
\name{plot_distributions_difference,successes_results-method}
\alias{plot_distributions_difference,successes_results-method}
\alias{plot_distributions_difference,ANY-method}
\title{plot_distributions_difference}
\usage{
\S4method{plot_distributions_difference}{successes_results}(object, ...)
}
\description{
\code{plot_distributions_difference} a visualization of the difference between the distribution of the first group and the second group.
}
