% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/successes_results_class.R
\docType{methods}
\name{plot_fit,successes_results-method}
\alias{plot_fit,successes_results-method}
\alias{plot_fit,ANY-method}
\title{plot_fit}
\usage{
\S4method{plot_fit}{successes_results}(object)
}
\description{
\code{plot_fit} plots fitted model against the data. Use this function to explore the quality of your fit.
}
