% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/color_class.R
\docType{methods}
\name{summary,color_class-method}
\alias{summary,color_class-method}
\title{summary}
\usage{
\S4method{summary}{color_class}(object)
}
\arguments{
\item{object}{color_class object.}
}
\description{
\code{summary} prints summary of the Bayesian color fit.
}
