% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linear_class.R
\docType{methods}
\name{summary,linear_class-method}
\alias{summary,linear_class-method}
\title{summary}
\usage{
\S4method{summary}{linear_class}(object)
}
\arguments{
\item{object}{linear_class object.}
}
\description{
\code{summary} prints summary of the Bayesian linear model fit.
}
