% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reaction_time_class.R
\docType{methods}
\name{summary,reaction_time_class-method}
\alias{summary,reaction_time_class-method}
\title{summary}
\usage{
\S4method{summary}{reaction_time_class}(object)
}
\arguments{
\item{object}{reaction_time_class object.}
}
\description{
\code{summary} prints summary of the Bayesian reaction time fit.
}
