% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/success_rate_class.R
\docType{methods}
\name{summary,success_rate_class-method}
\alias{summary,success_rate_class-method}
\title{summary}
\usage{
\S4method{summary}{success_rate_class}(object)
}
\arguments{
\item{object}{success_rate_class object.}
}
\description{
\code{summary} prints a summary of the Bayesian success rate fit.
}
