% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ttest_class.R
\docType{methods}
\name{summary,ttest_class-method}
\alias{summary,ttest_class-method}
\title{summary}
\usage{
\S4method{summary}{ttest_class}(object)
}
\arguments{
\item{object}{ttest_class object.}
}
\description{
\code{summary} prints summary of the Bayesian ttest fit.
}
