% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ttest_bayes.R
\name{ttest_bayes}
\alias{ttest_bayes}
\title{Bayesian t-test}
\usage{
ttest_bayes(y1, y2, ROPE = 0)
}
\arguments{
\item{y1}{Numeric vector of values for the first group.}

\item{ROPE}{optional parameter for region of practical equivalence, defaults = NULL.}

\item{y1}{Numeric vector of values for the second group.}

\item{seed}{Optional parameter for determinng random seed used in Stan MCMC, default = sample.int(.Machine$integer.max, 1)}
}
\value{
An object of class `ttest_bayes_results`
}
\description{
Bayesian t-test
}
