% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ttest_class.R
\name{compare_means,ttest_class-method}
\alias{compare_means,ttest_class-method}
\alias{compare_means_ttest}
\title{compare_means}
\usage{
\S4method{compare_means}{ttest_class}(object, ...)
}
\arguments{
\item{object}{ttest_class object.}

\item{...}{fit2 - a second ttest_class object, mu - mean value, sigma - standard deviation, fits - a list of ttest_class objects, rope - region of practical equivalence, par - execute comparison through the sigma or nu parameter.}
}
\value{
Comparison results or an error if something went wrong.
}
\description{
\code{compare_means} prints difference/equality of the first group against the second group, against multiple groups, against a mean value or against a normal distribution with a defined mean value and variance.
}
\examples{
# to use the function you first have to prepare the data and fit the model
# see class documentation for an example of the whole process
# along with an example of how to use this function
?ttest_class

}
