% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ttest_class.R
\docType{methods}
\name{compare_samples,ttest_class-method}
\alias{compare_samples,ttest_class-method}
\alias{compare_samples_ttest}
\title{compare_samples}
\usage{
\S4method{compare_samples}{ttest_class}(object, ...)
}
\arguments{
\item{object}{ttest_class object.}

\item{...}{fit2 - a second ttest_class object, mu - mean value, sigma - standard deviation, fits - a list of ttest_class objects, rope - region of practical equivalence.}
}
\description{
\code{compare_samples} prints difference/equality of the first group against the second group, against a mean value or against a normal distribution with a defined mean value and variance.
}
