% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ttest_class.R
\docType{methods}
\name{get_parameters,ttest_class-method}
\alias{get_parameters,ttest_class-method}
\alias{get_parameters_ttest_class}
\title{get_parameters}
\usage{
\S4method{get_parameters}{ttest_class}(object)
}
\arguments{
\item{object}{ttest_class object.}
}
\description{
\code{get_parameters} returns a dataframe with values of fitted parameters.
}
