% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ttest_class.R
\docType{methods}
\name{get_samples,ttest_class-method}
\alias{get_samples,ttest_class-method}
\alias{get_samples_ttest_class}
\title{get_samples}
\usage{
\S4method{get_samples}{ttest_class}(object)
}
\arguments{
\item{object}{ttest_class object.}
}
\description{
\code{get_samples} returns a dataframe with values of fitted parameters.
}
