% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ttest_class.R
\docType{methods}
\name{plot_comparison,ttest_class-method}
\alias{plot_comparison,ttest_class-method}
\alias{plot_comparison,ANY-method}
\title{plot_comparison}
\usage{
\S4method{plot_comparison}{ttest_class}(object, ...)
}
\description{
\code{plot_comparison} plots density for the first group and density for the second group, or a mean value in case second group is defined as a normal distribution or as a constant.
}
