% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ttest_class.R
\docType{methods}
\name{plot_distributions,ttest_class-method}
\alias{plot_distributions,ttest_class-method}
\alias{plot_distributions,ANY-method}
\title{plot_distributions}
\usage{
\S4method{plot_distributions}{ttest_class}(object, ...)
}
\description{
\code{plot_distributions} visualizes distributions underlying tested groups.
}
