% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ttest_class.R
\docType{methods}
\name{plot_fit,ttest_class-method}
\alias{plot_fit,ttest_class-method}
\alias{plot_fit_ttest}
\title{plot_fit}
\usage{
\S4method{plot_fit}{ttest_class}(object)
}
\arguments{
\item{object}{ttest_class object.}
}
\description{
\code{plot_fit} plots fitted model against the data. Use this function to explore the quality of your fit.
}
