% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ttest_class.R
\name{plot_means_difference,ttest_class-method}
\alias{plot_means_difference,ttest_class-method}
\alias{plot_means_difference_ttest}
\title{plot_means_difference}
\usage{
\S4method{plot_means_difference}{ttest_class}(object, ...)
}
\arguments{
\item{object}{ttest_class object.}

\item{...}{fit2 - a second ttest_class object, fits - a list of ttest_class objects, mu - mean value, rope - region of practical equivalence, bins - number of bins in the histogram, par - compare through the sigma or nu parameter.}
}
\value{
A ggplot visualization or an error if something went wrong.
}
\description{
\code{plot_means_difference} a visualization of the difference of the first group against the second group, against multiple groups, against a mean value or against a normal distribution with a defined mean value and variance.
}
\examples{
# to use the function you first have to prepare the data and fit the model
# see class documentation for an example of the whole process
# along with an example of how to use this function
?ttest_class

}
