% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ttest_class.R
\docType{methods}
\name{plot_samples_difference,ttest_class-method}
\alias{plot_samples_difference,ttest_class-method}
\alias{plot_samples_difference_ttest}
\title{plot_samples_difference}
\usage{
\S4method{plot_samples_difference}{ttest_class}(object, ...)
}
\arguments{
\item{object}{ttest_class object.}

\item{...}{fit2 - a second ttest_class object, fits - a list of ttest_class objects, mu - mean value, rope - region of practical equivalence, bins - number of bins in the histogram.}
}
\description{
\code{plot_samples_difference} a visualization of the difference of the first group against the second group, against multiple groups, against a mean value or against a normal distribution with a defined mean value and variance.
}
