% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ttest_class.R
\docType{methods}
\name{plot_trace,ttest_class-method}
\alias{plot_trace,ttest_class-method}
\alias{plot_trace_ttest}
\title{plot_trace}
\usage{
\S4method{plot_trace}{ttest_class}(object)
}
\arguments{
\item{object}{ttest_class object.}
}
\description{
\code{plot_trace} traceplot for main fitted model parameters.
}
