% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ttest_class.R
\docType{methods}
\name{traceplot,ttest_class-method}
\alias{traceplot,ttest_class-method}
\alias{traceplot,ANY-method}
\title{traceplot}
\usage{
\S4method{traceplot}{ttest_class}(object)
}
\description{
\code{traceplot} traceplot for main fitted model parameters.
}
