% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ttest_results_class.R
\docType{methods}
\name{compare,ttest_results-method}
\alias{compare,ttest_results-method}
\alias{compare,ANY-method}
\title{compare}
\usage{
\S4method{compare}{ttest_results}(object)
}
\description{
\code{compare} prints difference/equality of the first group against the second group, against a mean value, or against a normal distribution with a defined mean value and variance.
}
