% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ttest_results_class.R
\docType{methods}
\name{compare_distributions}
\alias{compare_distributions}
\alias{compare_distributions,ttest_results-method}
\alias{compare_distributions,ANY-method}
\title{compare_distributions}
\usage{
compare_distributions(object)

\S4method{compare_distributions}{ttest_results}(object)
}
\description{
\code{compare_distributions} prints difference/equality of samples drawn from fitted distributions.
}
