% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ttest_results_class.R
\docType{methods}
\name{comparison_plot}
\alias{comparison_plot}
\alias{comparison_plot,ttest_results-method}
\alias{comparison_plot,ANY-method}
\title{comparison_plot}
\usage{
comparison_plot(object)

\S4method{comparison_plot}{ttest_results}(object)
}
\description{
\code{comparison_plot} visualizes both tested groups.
}
