% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ttest_results_class.R
\docType{methods}
\name{difference_plot}
\alias{difference_plot}
\alias{difference_plot,ttest_results-method}
\alias{difference_plot,ANY-method}
\title{difference_plot}
\usage{
difference_plot(object)

\S4method{difference_plot}{ttest_results}(object)
}
\description{
\code{difference_plot} visualizes difference/equality of two tested groups.
}
