% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ttest_results_class.R
\docType{methods}
\name{plot_comparison}
\alias{plot_comparison}
\alias{plot_comparison,ttest_results-method}
\alias{plot_comparison,ANY-method}
\title{plot_comparison}
\usage{
plot_comparison(object)

\S4method{plot_comparison}{ttest_results}(object)
}
\description{
\code{plot_comparison} visualizes both tested groups.
}
