% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ttest_results_class.R
\docType{methods}
\name{plot_difference,ttest_results-method}
\alias{plot_difference,ttest_results-method}
\alias{plot_difference,ANY-method}
\title{plot_difference}
\usage{
\S4method{plot_difference}{ttest_results}(object, ..., bins = 30)
}
\description{
\code{plot_difference} a visualization of the difference of the first group against the second group, against a mean value, or against a normal distribution with a defined mean value and variance.
}
