% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ttest_results_class.R
\docType{methods}
\name{plot_distributions_difference}
\alias{plot_distributions_difference}
\alias{plot_distributions_difference,ttest_results-method}
\alias{plot_distributions_difference,ANY-method}
\title{plot_distributions_difference}
\usage{
plot_distributions_difference(object)

\S4method{plot_distributions_difference}{ttest_results}(object)
}
\description{
\code{plot_distributions_difference} visualizes difference/equality of distributions underlying the data.
}
