% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ttest_results_class.R
\docType{methods}
\name{traceplot,ttest_results-method}
\alias{traceplot,ttest_results-method}
\alias{traceplot,ANY-method}
\title{traceplot}
\usage{
\S4method{traceplot}{ttest_results}(object)
}
\description{
\code{traceplot} traceplot for main fitted model parameters.
}
