% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/b_color.R
\name{b_color}
\alias{b_color}
\title{b_color}
\usage{
b_color(colors, priors = NULL, hsv = FALSE, warmup = 2000,
  iter = 3000, chains = 4, control = NULL)
}
\arguments{
\item{colors}{a data frame of colors either in RGB or HSV format. The first column should be the R (or H) component, the second column should be the G (or S) component and the third columne should be the B (or V) component.}

\item{priors}{List of parameters and their priors - b_prior objects. You can put a prior on the mu_r (mean r component), sigma_r (variance of mu_r), mu_g (mean g component), sigma_g (variance of mu_g), mu_b (mean b component), sigma_b (variance of mu_b), mu_h (mean h component), kappa_h (variance of mu_h), mu_s (mean s component), sigma_s (variance of mu_s), mu_v (mean v component) and sigma_v (variance of mu_v) parameters (default = NULL).}

\item{hsv}{set to TRUE if colors are provided in HSV format (default = FALSE).}

\item{warmup}{Integer specifying the number of warmup iterations per chain (default = 2000).}

\item{iter}{Integer specifying the number of iterations (including warmup, default = 3000).}

\item{chains}{Integer specifying the number of parallel chains (default = 4).}

\item{control}{A named list of parameters to control the sampler's behavior (default = NULL).}
}
\value{
An object of class `color_class`
}
\description{
Bayesian model for comparing colors.
}
