% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/b_linear.R
\name{b_linear}
\alias{b_linear}
\title{b_linear}
\usage{
b_linear(x, y, s)
}
\arguments{
\item{x}{a vector containting index of sequence (time).}

\item{y}{a vector containting subjet's responses.}

\item{s}{a vector contaiting subject indexes. Starting index should be 1, and the largest subject index equals m (number of subjects).}
}
\value{
An object of class `linear_class`.
}
\description{
Bayesian model for fitting a linear normal model to data.
}
