% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/b_reaction_time.R
\name{b_reaction_time}
\alias{b_reaction_time}
\title{Bayesian model for comparing reaction times and test success rate.}
\usage{
b_reaction_time(n, m, rt, s)
}
\arguments{
\item{n}{total number of measurements.}

\item{m}{number of participating subjects.}

\item{rt}{a vector containing reaction times for each measurement.}

\item{s}{a vector contaiting subject indexes. Starting index should be 1, and the largest subject index equals m (number of subjects).}
}
\value{
An object of class `reaction_time_class`.
}
\description{
Bayesian model for comparing reaction times and test success rate.
}
