% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/b_results.R
\docType{class}
\name{b_results-class}
\alias{b_results-class}
\title{b_results}
\description{
Parent S4 class for declaring shared function generics.

\strong{Functions}

summary(`b_results_class`): prints summary of the fit.

print(`b_results_class`): prints a more detailed summary of the fit.

show(`b_results_class`): prints a more detailed summary of the fit.
}
