% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/b_results.R
\name{compare_distributions}
\alias{compare_distributions}
\title{compare_distributions}
\usage{
compare_distributions(object, ...)
}
\arguments{
\item{object}{S4 class object from bayes4psy library.}

\item{...}{see documentation for specific class for the description of available parameters, e.g. ?compare_distributions_ttest or ?compare_distributions_linear.}
}
\description{
\code{compare_distributions} draws samples from distribution of the first fit and compares them against samples drawn from the distribution of the second fit, or against samples from multiple fits.
}
