% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/b_results.R
\name{compare_means}
\alias{compare_means}
\title{compare_means}
\usage{
compare_means(object, ...)
}
\arguments{
\item{object}{S4 class object from bayes4psy library.}

\item{...}{see documentation for specific class for the description of available parameters, e.g. ?compare_means_ttest or ?compare_means_linear.}
}
\description{
\code{compare_means} prints difference between means of two or multiple fits.
}
