% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/b_results.R
\name{plot_distributions}
\alias{plot_distributions}
\title{plot_distributions}
\usage{
plot_distributions(object, ...)
}
\arguments{
\item{object}{S4 class object from EasyBayes library.}

\item{...}{see documentation for specific class for the description of available parameters, e.g. ?plot_distributions_ttest or ?plot_distributions_linear.}
}
\description{
\code{plot_distributions} visualizes distributions underlying tested groups.
}
