% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/b_results.R
\name{plot_distributions_difference}
\alias{plot_distributions_difference}
\title{plot_distributions_difference}
\usage{
plot_distributions_difference(object, ...)
}
\arguments{
\item{object}{S4 class object from EasyBayes library.}

\item{...}{see documentation for specific class for the description of available parameters.}
}
\description{
\code{plot_distributions_difference} a visualization of the difference between the distribution of the first group and the distribution or a constant value for the second group.
}
