% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/b_results.R
\name{plot_fit}
\alias{plot_fit}
\title{plot_fit}
\usage{
plot_fit(object, ...)
}
\arguments{
\item{object}{S4 class object from bayes4psy library.}

\item{...}{see documentation for specific class for the description of available parameters, e.g. ?plot_fit_colors or ?plot_fit_linear.}
}
\description{
\code{plot_fit} plots fitted model against the data. Use this function to explore the quality of your fit.
}
