% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/b_results.R
\name{plot_trace}
\alias{plot_trace}
\title{plot_trace}
\usage{
plot_trace(object)
}
\arguments{
\item{object}{S4 class object from bayes4psy library.}
}
\description{
\code{plot_trace} traceplot for main fitted model parameters.
}
