% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/b_ttest.R
\name{b_ttest}
\alias{b_ttest}
\title{b_ttest}
\usage{
b_ttest(y)
}
\arguments{
\item{y}{Numeric vector of values on which the fit will be based.}
}
\value{
An object of class `ttest_class`.
}
\description{
Bayesian t-test.
}
