% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/color_class.R
\name{plot_means_difference,color_class-method}
\alias{plot_means_difference,color_class-method}
\alias{plot_means_difference_color}
\title{plot_means_difference}
\usage{
\S4method{plot_means_difference}{color_class}(object, ...)
}
\arguments{
\item{object}{color_class object.}

\item{...}{fit2 - a second color_class object, rgb - color defined through rgb, hsv - color defined through rgb, rope - region of practical equivalence, bins - number of bins in the histogram, pars - components of comparison, a subset of (r, g, b, h, s, v).}
}
\value{
A ggplot visualization or an error if something went wrong.
}
\description{
\code{plot_means_difference} a visualization of the difference between two fits
}
\examples{
# to use the function you first have to prepare the data and fit the model
# see class documentation for an example of the whole process
# along with an example of how to use this function
?color_class

}
