% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EasyBayes-package.R
\docType{package}
\name{EasyBayes-package}
\alias{EasyBayes-package}
\alias{EasyBayes}
\title{The 'EasyBayes' package.}
\description{
A DESCRIPTION OF THE PACKAGE
}
\references{
Stan Development Team (NA). RStan: the R interface to Stan. R package version NA. http://mc-stan.org
}
