% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linear_class.R
\docType{class}
\name{linear_class-class}
\alias{linear_class-class}
\alias{linear_class}
\title{linear_class}
\description{
An S4 class for storing results of normal linear model.
summary(`linear_class`): prints summary od the fit.

compare(`linear_class`, fit2 = `linear_class`): prints difference in slope and intercept between two groups. You can also provide the rope parameter.

plot_difference(`linear_class`, fit2 = `linear_class`): a visualization of the difference between two groups. You can also provide the rope parameter.

plot_samples(`linear_class`): plots density for the first group samples.

plot_samples(`linear_class`, fit2 = `linear_class`): plots density for the first and the second group samples.

compare_distributions(`linear_class`, fit2 = `linear_class`): draws samples from distribution of the first group and compares them against samples drawn from the distribution of the second group.

plot_distributions(`linear_class`): a visualization of the distribution for the first group.

plot_distributions(`linear_class`, fit2 = `linear_class`): a visualization of the distribution for the first group and the second group.

plot_distributions_difference(`linear_class`, fit2 = `linear_class`): a visualization of the difference between the distribution of the first group and the second group. You can also provide the rope parameter.

plot_fit(`linear_class`): plots fitted model against the data. Use this function to explore the quality of your fit.

traceplot(`linear_class`): traceplot for main fitted model parameters.
}
\section{Slots}{

\describe{
\item{\code{extract}}{Extract from Stan fit.}

\item{\code{fit}}{Stan fit.}

\item{\code{data}}{Raw data for the tested group.}
}}

