% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linear_class.R
\docType{methods}
\name{compare,linear_class-method}
\alias{compare,linear_class-method}
\alias{compare,ANY-method}
\title{compare}
\usage{
\S4method{compare}{linear_class}(object, ...)
}
\description{
\code{compare} prints difference in intercept and slope between two groups.
}
