% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linear_class.R
\name{get_subject_parameters,linear_class-method}
\alias{get_subject_parameters,linear_class-method}
\alias{get_subject_parameters_linear_class}
\title{get_subject_parameters}
\usage{
\S4method{get_subject_parameters}{linear_class}(object)
}
\arguments{
\item{object}{linear_class object.}
}
\value{
A data frame with parameter values.
}
\description{
\code{get_subject_parameters} returns a dataframe with values of fitted parameters for each subject in the hierarchical model.
}
\examples{
# to use the function you first have to prepare the data and fit the model
# see class documentation for an example of the whole process
# along with an example of how to use this function
?linear_class

}
