% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linear_class.R
\docType{methods}
\name{plot_samples,linear_class-method}
\alias{plot_samples,linear_class-method}
\alias{plot_difference,ANY-method}
\title{plot_difference}
\usage{
\S4method{plot_samples}{linear_class}(object, ...)
}
\description{
\code{plot_difference} TODO

\code{plot_samples} plots first group samples, or the first and the second group samples.
}
