% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linear_class.R
\docType{methods}
\name{plot_distributions_difference,linear_class-method}
\alias{plot_distributions_difference,linear_class-method}
\alias{plot_distributions,ANY-method}
\title{plot_distributions}
\usage{
\S4method{plot_distributions_difference}{linear_class}(object, ...)
}
\description{
\code{plot_distributions} TODO

\code{plot_distributions_difference} TODO
}
