% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linear_class.R
\docType{methods}
\name{plot_distributions,linear_class-method}
\alias{plot_distributions,linear_class-method}
\alias{plot_distributions_linear}
\title{plot_distributions}
\usage{
\S4method{plot_distributions}{linear_class}(object, ...)
}
\arguments{
\item{object}{linear_class object.}

\item{...}{fit2 - a second linear_class object.}
}
\description{
\code{plot_distributions} plots distribution of the first group and optionally the distribution of the second group.
}
