% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linear_class.R
\name{plot_distributions,linear_class-method}
\alias{plot_distributions,linear_class-method}
\alias{plot_distributions_linear}
\title{plot_distributions}
\usage{
\S4method{plot_distributions}{linear_class}(object, ...)
}
\arguments{
\item{object}{linear_class object.}

\item{...}{fit2 - a second linear_class object.}
}
\value{
A ggplot visualization or an error if something went wrong.
}
\description{
\code{plot_distributions} a visualization of the fitted distribution, for one or two fits.
}
\examples{
# to use the function you first have to prepare the data and fit the model
# see class documentation for an example of the whole process
# along with an example of how to use this function
?linear_class

}
