% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linear_class.R
\name{plot_means,linear_class-method}
\alias{plot_means,linear_class-method}
\alias{plot_means_linear}
\title{plot_means}
\usage{
\S4method{plot_means}{linear_class}(object, ...)
}
\arguments{
\item{object}{linear_class object.}

\item{...}{fit2 - a second linear_class object, par - plot a specific parameter (slope or intercept).}
}
\value{
A ggplot visualization or an error if something went wrong.
}
\description{
\code{plot_means} plots means or the first and the second group means.
}
\examples{
# to use the function you first have to prepare the data and fit the model
# see class documentation for an example of the whole process
# along with an example of how to use this function
?linear_class

}
