% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linear_class.R
\name{plot_means_difference,linear_class-method}
\alias{plot_means_difference,linear_class-method}
\alias{plot_means_difference_linear}
\title{plot_means_difference}
\usage{
\S4method{plot_means_difference}{linear_class}(object, ...)
}
\arguments{
\item{object}{linear_class object.}

\item{...}{fit2 - a second linear_class object, par - specific parameter of comparison (slope or intercept), rope_intercept and rope_slope - regions of practical equivalence, bins - number of bins in the histogram.}
}
\value{
A ggplot visualization or an error if something went wrong.
}
\description{
\code{plot_means_difference} plots difference between two groups.
}
\examples{
# to use the function you first have to prepare the data and fit the model
# see class documentation for an example of the whole process
# along with an example of how to use this function
?linear_class

}
