% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linear_class.R
\name{plot_trace,linear_class-method}
\alias{plot_trace,linear_class-method}
\alias{plot_trace_linear}
\title{plot_trace}
\usage{
\S4method{plot_trace}{linear_class}(object)
}
\arguments{
\item{object}{linear_class object.}
}
\value{
A ggplot visualization.
}
\description{
\code{plot_trace} traceplot for main fitted model parameters.
}
\examples{
# to use the function you first have to prepare the data and fit the model
# see class documentation for an example of the whole process
# along with an example of how to use this function
?linear_class

}
