% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcmc_hdi.r
\name{mcmc_hdi}
\alias{mcmc_hdi}
\title{mcmc_hdi}
\usage{
mcmc_hdi(samples, cred_mass = 0.95)
}
\arguments{
\item{samples}{A vector of representative values from a probability distribution.}

\item{cred_mass}{A scalar between 0 and 1, indicating the mass within the credible interval that is to be estimated.}
}
\value{
A vector containing the limits of the HDI.
}
\description{
Highest density interval from a sample of representative values, estimated as shortest credible interval.
}
\author{
John Kruschke
}
