% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shared_functions.R
\name{mcmc_hdi}
\alias{mcmc_hdi}
\title{mcmc_hdi}
\usage{
mcmc_hdi(samples, cred_mass = 0.95)
}
\arguments{
\item{samples}{vector of samples.}

\item{cred_mass}{credibility mass that the interval should include (default = 0.95).}
}
\value{
Boundaries of the HDI.
}
\description{
A function for calculating the HDI of samples (high density interval).
}
\author{
John Kruschke
}
