% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normal.R
\name{normal_stan}
\alias{normal_stan}
\title{Bayesian normal model}
\usage{
normal_stan(y, ...)
}
\arguments{
\item{y}{Numberic vector of output values.}

\item{...}{Arguments passed to `rstan::sampling` (e.g. iter, chains).}

\item{x}{Numeric vector of input values.}
}
\value{
An object of class `stanfit` returned by `rstan::sampling`
}
\description{
Bayesian normal model
}
