% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normal_stan.R
\name{normal_stan}
\alias{normal_stan}
\title{Bayesian normal model}
\usage{
normal_stan(y, log_skewed = FALSE)
}
\arguments{
\item{y}{Numeric vector of values.}

\item{log_skewed}{Optional parameter for determing if data is log skewed, default = FALSE.}

\item{warmup}{Optional parameter for number of burnin, warmup steps, default = 1000.}

\item{iter}{Optional parameter for number of iterations (warmup + samples), default = 1500.}

\item{seed}{Optional parameter for determinng random seed used in Stan MCMC, default = sample.int(.Machine$integer.max, 1)}
}
\value{
An object of class `stanfit` returned by `rstan::sampling`
}
\description{
Bayesian normal model
}
