% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reaction_time_class.R
\docType{class}
\name{reaction_time_class-class}
\alias{reaction_time_class-class}
\alias{reaction_time_class}
\title{reaction_time_class}
\description{
An S4 class for storing results of reaction time Bayesian model.

\strong{Functions}

summary(`reaction_time_class`): prints summary od the fit.

print(`reaction_time_class`): prints a more detailed summary of the fit

show(`reaction_time_class`): prints a more detailed summary of the fit.

get_samples(`reaction_time_class`): returns a dataframe with values of fitted parameters.

get_subject_samples(`reaction_time_class`): returns a dataframe with values of fitted parameters for each subject in the hierarchical model.

compare_samples(`reaction_time_class`, fit2=`reaction_time_class`): returns difference in reaction times between two groups. You can also provide the rope parameter or execute the comparison only through a chosen parameter - mu or lambda.

compare_samples(`reaction_time_class`, fits=`list`): returns difference in reaction times between multiple groups. You can also provide the rope parameter.

plot_samples_difference(`reaction_time_class`, fit2=`reaction_time_class`): a visualization of the difference between two groups. You can also provide the rope and bins (number of bins in the histogram) parameters or visualize the comparison only through a chosen parameter - mu or lambda.

plot_samples_difference(`reaction_time_class`, fits=`list`): a visualization of the difference between multiple groups. You can also provide the rope and bins (number of bins in the histogram) parameters or visualize the comparison only through a chosen parameter - mu or lambda.

plot_samples(`reaction_time_class`): plots density of the samples samples. You can also visualize the density only for a chosen parameter - mu or lambda.

plot_samples(`reaction_time_class`, fit2=`reaction_time_class`): plots density for the first and the second group samples. You can also visualize the density only for a chosen parameter - mu or lambda.

plot_samples(`reaction_time_class`, fits=`list`): plots density for samples of multiple groups. You can also visualize the density only for a chosen parameter - mu or lambda.

compare_distributions(`reaction_time_class`, fit2=`reaction_time_class`): draws samples from distribution of the first group and compares them against samples drawn from the distribution of the second group. You can also provide the rope parameter.

compare_distributions(`reaction_time_class`, fits=`lists`): draws and compares samples from distributions of multiple groups. You can also provide the rope parameter.

plot_distributions(`reaction_time_class`): a visualization of the fitted distribution.

plot_distributions(`reaction_time_class`, fit2=`reaction_time_class`): a visualization of the distribution for two fits.

plot_distributions(`reaction_time_class`, fits=`list`): a visualization of the distribution for multiple fits.

plot_distributions_difference(`reaction_time_class`, fit2=`reaction_time_class`): a visualization of the difference between the distribution of the first group and the second group. You can also provide the rope and bins (number of bins in the histogram) parameters.

plot_distributions_difference(`reaction_time_class`, fits=`list`): a visualization of the difference between the distributions of multiple groups. You can also provide the rope and bins (number of bins in the histogram) parameters.

plot_fit(`reaction_time_class`): plots fitted model against the data. Use this function to explore the quality of your fit.

plot_trace(`reaction_time_class`): traceplot for main fitted model parameters.
}
\section{Slots}{

\describe{
\item{\code{extract}}{Extract from Stan fit.}

\item{\code{fit}}{Stan fit.}

\item{\code{data}}{Data on which the fit is based.}
}}

