% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reaction_time_class.R
\docType{methods}
\name{compare,reaction_time_class-method}
\alias{compare,reaction_time_class-method}
\title{compare}
\usage{
\S4method{compare}{reaction_time_class}(object, ...)
}
\arguments{
\item{object}{reaction_time_class object.}

\item{...}{fit2 - a second linear_class object, rope - region of practical equivalence, par - specific parameter of comparison - mu or lambda.}
}
\description{
\code{compare} prints difference in reaction times between two groups.
}
