% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reaction_time_class.R
\name{compare_distributions,reaction_time_class-method}
\alias{compare_distributions,reaction_time_class-method}
\alias{compare_distributions_reaction_time}
\title{compare_distributions}
\usage{
\S4method{compare_distributions}{reaction_time_class}(object, ...)
}
\arguments{
\item{object}{reaction_time_class object.}

\item{...}{fit2 - a second reaction_time_class object, fits - a list of reaction_time_class objects, rope - region of practical equivalence.}
}
\value{
Comparison results or an error if something went wrong.
}
\description{
\code{compare_distributions} draws samples from distribution of the first group and compares them against samples drawn from the distribution of the second group or from samples drawn from distributions of multiple groups.
}
\examples{
# to use the function you first have to prepare the data and fit the model
# see class documentation for an example of the whole process
# along with an example of how to use this function
?reaction_time_class

}
