% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reaction_time_class.R
\name{get_parameters,reaction_time_class-method}
\alias{get_parameters,reaction_time_class-method}
\alias{get_parameters_reaction_time}
\title{get_parameters}
\usage{
\S4method{get_parameters}{reaction_time_class}(object)
}
\arguments{
\item{object}{reaction_time_class object.}
}
\value{
A data frame with parameter values.
}
\description{
\code{get_parameters} returns a dataframe with values of fitted parameters.
}
\examples{
# to use the function you first have to prepare the data and fit the model
# see class documentation for an example of the whole process
# along with an example of how to use this function
?reaction_time_class

}
