% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reaction_time_class.R
\name{plot_distributions,reaction_time_class-method}
\alias{plot_distributions,reaction_time_class-method}
\alias{plot_distributions_reaction_time}
\title{plot_distributions}
\usage{
\S4method{plot_distributions}{reaction_time_class}(object, ...)
}
\arguments{
\item{object}{reaction_time_class object.}

\item{...}{fit2 - a second reaction_time_class object, fits - a list of reaction_time_class objects.}
}
\value{
A ggplot visualization or an error if something went wrong.
}
\description{
\code{plot_distributions} a visualization of the fitted distribution, for one, two or multiple fits.
}
\examples{
# to use the function you first have to prepare the data and fit the model
# see class documentation for an example of the whole process
# along with an example of how to use this function
?reaction_time_class

}
