% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reaction_time_class.R
\name{plot_trace,reaction_time_class-method}
\alias{plot_trace,reaction_time_class-method}
\alias{plot_trace_reaction_time}
\title{plot_trace}
\usage{
\S4method{plot_trace}{reaction_time_class}(object)
}
\arguments{
\item{object}{reaction_time_class object.}
}
\value{
A ggplot visualization.
}
\description{
\code{plot_trace} traceplot for main fitted model parameters.
}
\examples{
# to use the function you first have to prepare the data and fit the model
# see class documentation for an example of the whole process
# along with an example of how to use this function
?reaction_time_class

}
