% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reaction_times_results_class.R
\docType{class}
\name{reaction_times_results-class}
\alias{reaction_times_results-class}
\alias{reaction_times_results}
\title{An S4 class for storing results of reaction time Bayesian model.}
\description{
An S4 class for storing results of reaction time Bayesian model.
}
\section{Slots}{

\describe{
\item{\code{extract}}{Extract from Stan fit.}

\item{\code{data}}{Data on which the fit is based.}
}}

\examples{
summary(`reaction_times_results`): prints summary od the fit.

plot_fit(`reaction_times_results`): plots fitted model against the data. Use this function to explore the quality of your fit.

plot_difference(`reaction_times_results`, `reaction_times_results`): a visualization of the difference between two groups.

}
