% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reaction_times_results_class.R
\docType{methods}
\name{plot_difference,reaction_times_results-method}
\alias{plot_difference,reaction_times_results-method}
\alias{plot_difference,ANY-method}
\title{plot_difference}
\usage{
\S4method{plot_difference}{reaction_times_results}(object, ...)
}
\description{
\code{plot_difference} visualizes difference/equality of two tested groups.
}
