% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reaction_times_results_class.R
\docType{methods}
\name{plot_fit}
\alias{plot_fit}
\alias{plot_fit,reaction_times_results-method}
\alias{plot_fit,ANY-method}
\title{plot_fit}
\usage{
plot_fit(object)

\S4method{plot_fit}{reaction_times_results}(object)
}
\description{
\code{plot_fit} visualizes fitted model against the data.
}
