% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reaction_times_results_class.R,
%   R/successes_results_class.R
\docType{methods}
\name{traceplot,reaction_times_results-method}
\alias{traceplot,reaction_times_results-method}
\alias{traceplot,ANY-method}
\alias{traceplot,successes_results-method}
\title{traceplot}
\usage{
\S4method{traceplot}{reaction_times_results}(object)

\S4method{traceplot}{successes_results}(object)
}
\description{
\code{traceplot} traceplot for main fitted model parameters.

\code{traceplot} traceplot for main fitted model parameters.
}
