% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shared_functions.R
\name{shared_difference}
\alias{shared_difference}
\title{shared_difference}
\usage{
shared_difference(y1, y2, rope = NULL)
}
\arguments{
\item{y1}{Numeric vector of values for the first group.}

\item{y2}{Numeric vector of values for the second group.}

\item{rope}{Optional parameter for the rope interval.}
}
\value{
Prints the difference and 95% HDI into the console.
}
\description{
A shared function for all classes for printing difference between two vectors of values.
}
