% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shared_plots.R
\name{shared_plot_difference}
\alias{shared_plot_difference}
\title{Shared plots used in several models.}
\usage{
shared_plot_difference(y1, y2, rope = NULL, bins = 30)
}
\arguments{
\item{y1}{Numeric vector of values for the first group.}

\item{y2}{Numeric vector of values for the second group.}

\item{rope}{Optional parameter for the rope interval.}
}
\value{
`ggplot` visualization of the difference.
}
\description{
Shared plots used in several models.
}
