% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linear_class.R
\name{show,linear_class-method}
\alias{show,linear_class-method}
\title{show}
\usage{
\S4method{show}{linear_class}(object)
}
\arguments{
\item{object}{linear_class object.}
}
\description{
\code{show} prints a more detailed summary of the Bayesian linear model fit.
}
\examples{
# to use the function you first have to prepare the data and fit the model
# see class documentation for an example of the whole process
# along with an example of how to use this function
?linear_class

}
