% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reaction_time_class.R
\name{show,reaction_time_class-method}
\alias{show,reaction_time_class-method}
\title{show}
\usage{
\S4method{show}{reaction_time_class}(object)
}
\arguments{
\item{object}{reaction_time_class object.}
}
\description{
\code{show} prints a more detailed summary of the Bayesian reaction time fit.
}
\examples{
# to use the function you first have to prepare the data and fit the model
# see class documentation for an example of the whole process
# along with an example of how to use this function
?reaction_time_class

}
