% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/success_rate_class.R
\name{show,success_rate_class-method}
\alias{show,success_rate_class-method}
\title{show}
\usage{
\S4method{show}{success_rate_class}(object)
}
\arguments{
\item{object}{success_rate_class object.}
}
\description{
\code{show} prints a more detailed summary of the Bayesian success rate fit.
}
\examples{
# to use the function you first have to prepare the data and fit the model
# see class documentation for an example of the whole process
# along with an example of how to use this function
?success_rate_class

}
