% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/success_rate_class.R
\docType{class}
\name{success_rate_class-class}
\alias{success_rate_class-class}
\alias{success_rate_class}
\title{success_rate_class}
\description{
An S4 class for storing results of successes (true/false) Bayesian model.
}
\section{Slots}{

\describe{
\item{\code{extract}}{Extract from Stan fit.}

\item{\code{fit}}{Stan fit.}

\item{\code{data}}{Data on which the fit is based.}
}}

\examples{
summary(`success_rate_class`): prints summary od the fit.

compare(`success_rate_class`, fit2 = `success_rate_class`): prints difference in successfulness of two groups. You can also provide the rope parameter.

plot_difference(`success_rate_class`, fit2 = `success_rate_class`): a visualization of the difference between two groups. You can also provide the rope parameter.

plot_samples(`success_rate_class`): plots density for the first samples.

plot_samples(`success_rate_class`, fit2 = `success_rate_class`): plots density for the first and the second group samples.

compare_distributions(`success_rate_class`, fit2 = `success_rate_class`): draws samples from distribution of the first group and compares them against samples drawn from the distribution of the second group.

plot_distributions(`success_rate_class`, fit2 = `success_rate_class`): a visualization of the distribution for the first group and the second group.

plot_distributions_difference(`success_rate_class`, fit2 = `success_rate_class`): a visualization of the difference between the distribution of the first group and the second group. You can also provide the rope parameter.

plot_fit(`success_rate_class`): plots fitted model against the data. Use this function to explore the quality of your fit.

traceplot(`success_rate_class`): traceplot for main fitted model parameters.

}
