% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/success_rate_class.R
\docType{methods}
\name{plot_distributions,success_rate_class-method}
\alias{plot_distributions,success_rate_class-method}
\alias{plot_distributions_success_rate}
\title{plot_distributions}
\usage{
\S4method{plot_distributions}{success_rate_class}(object, ...)
}
\arguments{
\item{object}{success_rate_class object.}

\item{...}{fit2 - a second success_rate_class object.}
}
\description{
\code{plot_distributions} a visualization of the distribution for the first group, or the first group and the second group.
}
