% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/success_rate_class.R
\docType{methods}
\name{plot_distributions_difference,success_rate_class-method}
\alias{plot_distributions_difference,success_rate_class-method}
\alias{plot_distributions_difference,success_rate_class-method,ANY-method}
\title{plot_distributions_difference}
\usage{
\S4method{plot_distributions_difference}{success_rate_class}(object, ...)
}
\arguments{
\item{object}{success_rate_class object.}

\item{...}{fit2 - a second success_rate_class object.}
}
\description{
\code{plot_distributions_difference} a visualization of the difference between the distribution of the first group and the second group.
}
