% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/success_rate_class.R
\docType{methods}
\name{plot_fit,success_rate_class-method}
\alias{plot_fit,success_rate_class-method}
\alias{plot_fit,success_rate_class-method,ANY-method}
\title{plot_fit}
\usage{
\S4method{plot_fit}{success_rate_class}(object)
}
\arguments{
\item{object}{success_rate_class object.}
}
\description{
\code{plot_fit} plots fitted model against the data. Use this function to explore the quality of your fit.
}
