% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/success_rate_class.R
\name{plot_means,success_rate_class-method}
\alias{plot_means,success_rate_class-method}
\alias{plot_means_success_rate}
\title{plot_means}
\usage{
\S4method{plot_means}{success_rate_class}(object, ...)
}
\arguments{
\item{object}{success_rate_class object.}

\item{...}{fit2 - a second success_rate_class object, fits - a list of success_rate_class objects.}
}
\value{
A ggplot visualization or an error if something went wrong.
}
\description{
\code{plot_means} plots density of means for one, two or multiple groups.
}
\examples{
# to use the function you first have to prepare the data and fit the model
# see class documentation for an example of the whole process
# along with an example of how to use this function
?success_rate_class

}
