% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/color_class.R
\name{summary,color_class-method}
\alias{summary,color_class-method}
\title{summary}
\usage{
\S4method{summary}{color_class}(object)
}
\arguments{
\item{object}{color_class object.}
}
\description{
\code{summary} prints summary of the Bayesian color fit.
}
\examples{
# to use the function you first have to prepare the data and fit the model
# see class documentation for an example of the whole process
# along with an example of how to use this function
?color_class

}
