% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ttest_class.R
\name{summary,ttest_class-method}
\alias{summary,ttest_class-method}
\title{summary}
\usage{
\S4method{summary}{ttest_class}(object)
}
\arguments{
\item{object}{ttest_class object.}
}
\description{
\code{summary} prints a summary of the Bayesian ttest fit.
}
\examples{
# to use the function you first have to prepare the data and fit the model
# see class documentation for an example of the whole process
# along with an example of how to use this function
?ttest_class

}
