% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ttest_bayes.R
\name{ttest_bayes}
\alias{ttest_bayes}
\title{Bayesian t-test.}
\usage{
ttest_bayes(y1, y2 = NULL, mu = NULL, sigma = NULL, rope = NULL)
}
\arguments{
\item{y1}{Numeric vector of values for the first group.}

\item{y2}{Optional parameter for values for the second group.}

\item{mu}{Optional parameter for comparison of first group against a mean value.}

\item{sigma}{Optional parameter for comparison of first group against a distribution N(mu, sigma).}

\item{rope}{Optional parameter for region of practical equivalence, default = NULL}
}
\value{
An object of class `ttest_bayes_results`.
}
\description{
Bayesian t-test.
}
