% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ttest_class.R
\name{plot_distributions,ttest_class-method}
\alias{plot_distributions,ttest_class-method}
\alias{plot_distributions_ttest}
\title{plot_distributions}
\usage{
\S4method{plot_distributions}{ttest_class}(object, ...)
}
\arguments{
\item{object}{ttest_class object.}

\item{...}{fit2 - a second ttest_class object, fits - a list of ttest_class objects, mu - mean value, sigma - standard deviation.}
}
\value{
A ggplot visualization or an error if something went wrong.
}
\description{
\code{plot_distributions} visualizes distributions underlying tested groups.
}
\examples{
# to use the function you first have to prepare the data and fit the model
# see class documentation for an example of the whole process
# along with an example of how to use this function
?ttest_class

}
