% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ttest_class.R
\name{plot_fit,ttest_class-method}
\alias{plot_fit,ttest_class-method}
\alias{plot_fit_ttest}
\title{plot_fit}
\usage{
\S4method{plot_fit}{ttest_class}(object)
}
\arguments{
\item{object}{ttest_class object.}
}
\value{
A ggplot visualization.
}
\description{
\code{plot_fit} plots fitted model against the data. Use this function to explore the quality of your fit.
}
\examples{
# to use the function you first have to prepare the data and fit the model
# see class documentation for an example of the whole process
# along with an example of how to use this function
?ttest_class

}
