% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ttest_fit.R
\name{ttest_fit}
\alias{ttest_fit}
\title{Bayesian Student's t-test model}
\usage{
ttest_fit(y)
}
\arguments{
\item{y}{Numeric vector of values.}
}
\value{
An object of class `stanfit` returned by `rstan::sampling`.
}
\description{
Bayesian Student's t-test model
}
