% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ttest_results_class.R
\docType{class}
\name{ttest_results-class}
\alias{ttest_results-class}
\alias{ttest_results}
\title{An S4 class for storing results of Bayesian t-test results.}
\description{
An S4 class for storing results of Bayesian t-test results.
}
\section{Slots}{

\describe{
\item{\code{y1_samples}}{samples for the first group.}

\item{\code{y2_samples}}{Optional variable: samples for the second group.}

\item{\code{mu}}{Optional variable: mean value for comparison with first group.}

\item{\code{sigma}}{Optional variable: variance for comparison with first group, used in combination with `mu` parameter.}

\item{\code{rope}}{Optional variable: region of practical equivalence.}
}}

\examples{
ttest_results: prints difference/equality of the first group against the second group, against a mean value, or against a normal distribution with a defined mean value and variance.

summary(ttest_results): prints difference/equality of the first group against the second group, against a mean value, or against a normal distribution with a defined mean value and variance.

plot_difference(ttest_results): a visualization of the difference of the first group against the second group, against a mean value, or against a normal distribution with a defined mean value and variance.

plot_comparison(ttest_results): plots histogram for the first group and the histogram of the second group, or a mean value in case second group is defined as a normal distribution or as a constant.

compare_distributions(ttest_results): draws samples from distribution of the first group and compares them against samples from the second group, against a mean value, or against samples from a normal distribution with a defined mean value and variance.

plot_distributions(ttest_results): a visualization of the distribution for the first group and the distribution or a constant value for the second group.

plot_distributions_difference(ttest_results): a visualization of the difference between the distribution of the first group and the distribution or a constant value for the second group.

}
