% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ttest_results_class.R
\docType{class}
\name{ttest_results-class}
\alias{ttest_results-class}
\alias{ttest_results}
\title{An S4 class for storing results of Bayesian t-test results.}
\description{
An S4 class for storing results of Bayesian t-test results.
}
\section{Slots}{

\describe{
\item{\code{y1_samples}}{samples for the first group.}

\item{\code{y2_samples}}{samples for the second group.}

\item{\code{ROPE}}{region of practical equivalence.}
}}

\examples{
ttest_results: prints difference/equality of two tested groups.

summary(ttest_resuolts): prints difference/equality of two tested groups.

difference_plot(ttest_results): visaualizes difference between two groups.
}
