% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ttest_results_class.R
\docType{methods}
\name{plot_difference}
\alias{plot_difference}
\alias{plot_difference,ttest_results-method}
\alias{plot_difference,ANY-method}
\title{plot_difference}
\usage{
plot_difference(object)

\S4method{plot_difference}{ttest_results}(object)
}
\description{
\code{plot_difference} visualizes difference/equality of two tested groups.
}
